Better pruning of events from Pythia8 and enable in HepMC
This commit
- Fixes the pruning of events in `GeneratorPythia8`.  The old
  implementation had a number of problems

  - Beam particles (HepMC status = 4) was _not_ kept
  - The event structure was _not_ preserved.  That meant that the event
    would be split into many disjoint trees.  Thus, subsequent passes
    over the event would not be able to tie things together.  This was
    especially a problem when converting the event to HepMC (via AOD)
    since the events would not be complete.
- Added the option to prune events read in via `GeneratorHepMC`.  The
  configuration key `HepMC.prune=true` will enable this mode (off by
  default).

By _pruning_, I mean that all particles that are _not_
- beam particles (HepMC status = 4),
- decayed (HepMC status = 2), nor
- final state (HepMC status = 1)
are removed from the event.  When we do that, it is extremely important
to connect up particles so that we have a consistent event structure.
For example, we may remove a gluon from the tree.  In that case, we must
take care to move all child particles to the proper node in the tree, or
the children will be disjoint from the rest of the event.

This patch also adds Doxygen documentation to `GeneratorPythia8`, as
well as a new `run/SimExamples/Pythia8/README.md` file.